requirements.txt
absl-py==0.7.1
appdirs==1.4.3
astor==0.8.0
beautifulsoup4==4.8.0
bleach==1.5.0
Click==7.0
cmake==3.14.4
cycler==0.10.0
decorator==4.4.0
dlib==19.17.0
enum34==1.1.6
gast==0.2.2
google-images-download==2.8.0
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
html5lib==0.9999999
imutils==0.5.2
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-vggface==0.6
kiwisolver==1.1.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.16.2
opencv-contrib-python==4.1.0.25
opencv-python==4.1.0.25
Pillow==6.1.0
protobuf==3.9.1
pycuda==2019.1.2
pyparsing==2.4.2
python-dateutil==2.8.0
python-speech-features==0.6
pytools==2019.1.1
PyYAML==5.1.2
scenedetect==0.5.1.1
scikit-learn==0.21.3
scipy==1.3.1
selenium==3.141.0
six==1.12.0
sklearn==0.0
soupsieve==1.9.2
tensorboard==1.12.2
tensorflow-gpu==1.12.0
termcolor==1.1.0
torch==1.2.0
urllib3==1.25.3
utils==0.9.0
Werkzeug==0.15.5
wrapt==1.11.2
youtube-dl==2019.8.13
